+++ NOLF2/ClientShellDLL/ClientShellShared/HUDMessageQueue.cpp
@@ -294,15 +294,15 @@ void CHUDMessageQueue::SetHistoryOffset(uint16 nOffset)
 	if (m_History.size() <= m_HistoryMsgs.size())
 		nMaxOffset = 0;
 	else
-		nMaxOffset = (m_History.size() - m_HistoryMsgs.size());
+		nMaxOffset = (uint16)(m_History.size() - m_HistoryMsgs.size());
 	
 	if (nOffset > nMaxOffset)
 		nOffset = nMaxOffset;
 
 	m_nHistoryOffset = nOffset;
 
-	uint8 nMsg = m_HistoryMsgs.size() - 1;
-	uint8 nIndex = (m_History.size() - m_nHistoryOffset) - 1;
+	uint8 nMsg = (uint8)(m_HistoryMsgs.size() - 1);
+	uint8 nIndex = (uint8)((m_History.size() - m_nHistoryOffset) - 1);
 
 	while (nMsg < m_HistoryMsgs.size())
 	{
@@ -332,9 +332,9 @@ void CHUDMessageQueue::IncHistoryOffset()
 	if (m_History.size() <= m_HistoryMsgs.size())
 		nMaxOffset = 0;
 	else
-		nMaxOffset = (m_History.size() - m_HistoryMsgs.size());
+		nMaxOffset = (uint16)(m_History.size() - m_HistoryMsgs.size());
 
-	uint16 nOffset = m_nHistoryOffset + m_HistoryMsgs.size();
+	uint16 nOffset = (uint16)(m_nHistoryOffset + m_HistoryMsgs.size());
 
 	if (nOffset > nMaxOffset)
 		nOffset = nMaxOffset;
@@ -348,7 +348,7 @@ void CHUDMessageQueue::DecHistoryOffset()
 	uint16 nOffset = m_nHistoryOffset;
 
 	if (nOffset > m_HistoryMsgs.size())
-		nOffset -= m_HistoryMsgs.size();
+		nOffset -= (uint16)(m_HistoryMsgs.size());
 	else
 		nOffset = 0;
