+++ NOLF2/ClientShellDLL/ClientShellShared/TimedText.cpp
@@ -78,7 +78,7 @@ void CTimedText::Init( TIMED_TEXT_INIT_STRUCT const &initData )
 		CLEARFLAG( m_flags, CTIMEDTEXTFLAG_USE_CLIP_RECT );
 	}
 	
-	m_textLen = strlen( m_text->GetText() );
+	m_textLen = (int)strlen( m_text->GetText() );
 	m_numberOfLinesBeforeScroll = initData.numberOfLinesBeforeScroll;
 		
 	// Timing Information
